# Faster trie node encode