--- a/trie/fast_node_encoder.go
+++ b/trie/fast_node_encoder.go
@@ -6,87 +6,66 @@
 package trie
 
 import (
-	"io"
-
 	"github.com/vechain/thor/lowrlp"
 )
 
-// fastNodeEncoder is the fast node encoder using low-level rlp encoder.
-type fastNodeEncoder struct{}
+// implements node.encode and node.encodeTrailing
 
-var frlp fastNodeEncoder
-
-// Encode writes the RLP encoding of node to w.
-func (fastNodeEncoder) Encode(w io.Writer, node node, nonCrypto bool) error {
-	enc := lowrlp.NewEncoder()
-	defer enc.Release()
-
-	fastEncodeNode(enc, node, nonCrypto)
-	return enc.ToWriter(w)
-}
-
-// EncodeToBytes returns the RLP encoding of node.
-func (fastNodeEncoder) EncodeToBytes(collapsed node, nonCrypto bool) []byte {
-	enc := lowrlp.NewEncoder()
-	defer enc.Release()
-
-	fastEncodeNode(enc, collapsed, nonCrypto)
-	return enc.ToBytes()
-}
-
-func (fastNodeEncoder) EncodeTrailing(w io.Writer, collapsed node) error {
-	enc := lowrlp.NewEncoder()
-	defer enc.Release()
-	fastEncodeNodeTrailing(enc, collapsed)
-	return enc.ToWriter(w)
-}
-
-func fastEncodeNode(w *lowrlp.Encoder, collapsed node, nonCrypto bool) {
-	switch n := collapsed.(type) {
-	case *fullNode:
-		offset := w.List()
-		for _, c := range n.Children {
-			if c != nil {
-				fastEncodeNode(w, c, nonCrypto)
-			} else {
-				w.EncodeEmptyString()
-			}
-		}
-		w.ListEnd(offset)
-	case *shortNode:
-		offset := w.List()
-		w.EncodeString(n.Key)
-		if n.Val != nil {
-			fastEncodeNode(w, n.Val, nonCrypto)
+func (n *fullNode) encode(e *lowrlp.Encoder, nonCrypto bool) {
+	off := e.List()
+	for _, c := range n.Children {
+		if c != nil {
+			c.encode(e, nonCrypto)
 		} else {
-			w.EncodeEmptyString()
+			e.EncodeEmptyString()
 		}
-		w.ListEnd(offset)
-	case *hashNode:
-		if nonCrypto {
-			w.EncodeString(nonCryptoNodeHashPlaceholder)
-		} else {
-			w.EncodeString(n.Hash[:])
+	}
+	e.ListEnd(off)
+}
+
+func (n *fullNode) encodeTrailing(e *lowrlp.Encoder) {
+	for _, c := range n.Children {
+		if c != nil {
+			c.encodeTrailing(e)
 		}
-	case *valueNode:
-		w.EncodeString(n.Value)
 	}
 }
 
-func fastEncodeNodeTrailing(w *lowrlp.Encoder, collapsed node) {
-	switch n := collapsed.(type) {
-	case *shortNode:
-		fastEncodeNodeTrailing(w, n.Val)
-	case *fullNode:
-		for _, c := range n.Children {
-			fastEncodeNodeTrailing(w, c)
-		}
-	case *valueNode:
-		// skip empty node
-		if len(n.Value) > 0 {
-			w.EncodeString(n.meta)
-		}
-	case *hashNode:
-		w.EncodeUint(n.seq)
+func (n *shortNode) encode(e *lowrlp.Encoder, nonCrypto bool) {
+	off := e.List()
+	e.EncodeString(n.Key)
+	if n.Val != nil {
+		n.Val.encode(e, nonCrypto)
+	} else {
+		e.EncodeEmptyString()
+	}
+	e.ListEnd(off)
+}
+
+func (n *shortNode) encodeTrailing(e *lowrlp.Encoder) {
+	if n.Val != nil {
+		n.Val.encodeTrailing(e)
+	}
+}
+
+func (n *hashNode) encode(e *lowrlp.Encoder, nonCrypto bool) {
+	if nonCrypto {
+		e.EncodeString(nonCryptoNodeHashPlaceholder)
+	} else {
+		e.EncodeString(n.Hash[:])
+	}
+}
+
+func (n *hashNode) encodeTrailing(e *lowrlp.Encoder) {
+	e.EncodeUint(n.seq)
+}
+
+func (n *valueNode) encode(e *lowrlp.Encoder, nonCrypto bool) {
+	e.EncodeString(n.Value)
+}
+
+func (n *valueNode) encodeTrailing(e *lowrlp.Encoder) {
+	if len(n.Value) > 0 {
+		e.EncodeString(n.meta)
 	}
 }