In this PR, lowrlp.Encoder is not pooled, instead the encoder is cached in hasher object( saves pool.Get/Put callings).

benchmarks:

```bash
$ benchstat before.txt after.txt 
name                  old time/op  new time/op  delta
FastEncodeFullNode-8   216ns ± 1%   180ns ± 0%  -16.97%  (p=0.008 n=5+5)
```